Make HTTP status code consistently a short
It used to be a short in some places and an int in other places.

FossilOrigin-Name: 6e42ee482ff02f76c845f6461688c20c318af1af9b3aa1802bab5103b334a969

+++ src/OFHTTPClient.h
@@ -101,7 +101,7 @@ OF_ASSUME_NONNULL_BEGIN
  */
 -      (void)client: (OFHTTPClient *)client
   didReceiveHeaders: (OFDictionary OF_GENERIC(OFString *, OFString *) *)headers
-	 statusCode: (int)statusCode
+	 statusCode: (short)statusCode
 	    request: (OFHTTPRequest *)request;
 
 /**
@@ -130,7 +130,7 @@ OF_ASSUME_NONNULL_BEGIN
  */
 -	  (bool)client: (OFHTTPClient *)client
   shouldFollowRedirect: (OFURL *)URL
-	    statusCode: (int)statusCode
+	    statusCode: (short)statusCode
 	       request: (OFHTTPRequest *)request
 	      response: (OFHTTPResponse *)response;
 @end
